requirements.txt
scapy
python-nmap
requests
flask
fpdf2
